endpoints for editing article and adding/removing fotos

+++ src/pipes/image-resize.pipe.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotAcceptableException, PipeTransform } from '@nestjs/common';
+import { Injectable, NotAcceptableException, PipeTransform, SetMetadata } from '@nestjs/common';
 import { readFile, unlink } from 'fs/promises';
 import * as path from 'path';
 import { serverDataConf } from 'server.config';
@@ -8,40 +8,50 @@ import { UploadeFileMulter } from 'src/types/multer.type';
 import { checkFreeSpace } from 'src/utils/free-space-counter';
 
 @Injectable()
-export class SharpPipe implements PipeTransform<UploadeFileMulter, Promise<UploadeFileMulter>> {
+export class SharpPipe implements PipeTransform<UploadeFileMulter, Promise<UploadeFileMulter | null>> {
 
-    async transform(files: UploadeFileMulter): Promise<UploadeFileMulter> {
+    async transform(files: UploadeFileMulter): Promise<UploadeFileMulter | null> {
 
-        await Promise.all(
-            files.file.map(async (file) => {
 
-                if (file.size > serverDataConf.maxFotoSize) {
-                    await unlink(path.join(destionation(), file.filename));
-                    throw new NotAcceptableException(Error, 'Plik przekracza dozowolone 20 Mb');
-                };
+        try {
+            if (files.file) {
+                await Promise.all(
+                    files.file.map(async (file) => {
 
-                if (file.mimetype !== 'image/jpeg' && file.mimetype !== 'image/png') {
-                    await unlink(path.join(destionation(), file.filename));
-                    throw new NotAcceptableException(Error, 'Niedozwolony typ pliku');
-                };
+                        if (file.size > serverDataConf.maxFotoSize) {
+                            await unlink(path.join(destionation(), file.filename));
+                            throw new NotAcceptableException(Error, 'Plik przekracza dozowolone 20 Mb');
+                        };
 
-                const takenSpace = await checkFreeSpace();
+                        if (file.mimetype !== 'image/jpeg' && file.mimetype !== 'image/png') {
+                            await unlink(path.join(destionation(), file.filename));
+                            throw new NotAcceptableException(Error, 'Niedozwolony typ pliku');
+                        };
 
-                if (file.size + takenSpace > serverDataConf.maxServerSpace) {
-                    await unlink(path.join(destionation(), file.filename));
-                    throw new NotAcceptableException(Error, 'Przekroczono ilość dozwolonego miejsca');
-                };
+                        const takenSpace = await checkFreeSpace();
 
-                const readed = await readFile(path.join(destionation(), file.filename));
-                const filename = file.filename + '.webp';
+                        if (file.size + takenSpace > serverDataConf.maxServerSpace) {
+                            await unlink(path.join(destionation(), file.filename));
+                            throw new NotAcceptableException(Error, 'Przekroczono ilość dozwolonego miejsca');
+                        };
 
-                await sharp(readed)
-                    .resize(800)
-                    .webp({ effort: 3 })
-                    .toFile(path.join(destionation(), filename));
-            })
-        );
+                        const readed = await readFile(path.join(destionation(), file.filename));
+                        const filename = file.filename + '.webp';
 
-        return files;
+                        await sharp(readed)
+                            .resize(800)
+                            .webp({ effort: 3 })
+                            .toFile(path.join(destionation(), filename));
+                    })
+                );
+
+                return files;
+
+            } else return null;
+
+        } catch (err) {
+            console.log(err);
+            throw new NotAcceptableException(Error, 'Nieznany błąd serwera');
+        }
     };
 }
